@@ -21,10 +21,11 @@
     "tslint": "tslint -p tsconfig.json './src/**/*.ts'",
     "typecheck": "tsc -p tsconfig.json",
     "build": "parcel build --out-file dist/java-imports-search.js src/java-imports-search.ts",
-    "serve": "parcel serve --no-hmr --out-file dist/java-imports-search.js src/java-imports-search.ts",
+    "serve": "npm run symlink-package && parcel serve --no-hmr --out-file dist/java-imports-search.js src/java-imports-search.ts",
     "watch:typecheck": "tsc -p tsconfig.json -w",
     "watch:build": "tsc -p tsconfig.dist.json -w",
-    "sourcegraph:prepublish": "npm run build"
+    "sourcegraph:prepublish": "npm run build",
+    "symlink-package": "mkdirp dist && lnfs ./package.json ./dist/package.json"
   },
   "browserslist": [
     "last 1 Chrome versions",
@@ -33,11 +34,17 @@
     "last 1 Safari versions"
   ],
   "devDependencies": {
-    "@sourcegraph/tsconfig": "^3.0.0",
-    "@sourcegraph/tslint-config": "^12.0.0",
+    "@sourcegraph/tsconfig": "^4.0.0",
+    "@sourcegraph/tslint-config": "^13.4.0",
+    "lnfs-cli": "^2.1.0",
+    "mkdirp": "^0.5.1",
     "parcel-bundler": "^1.10.3",
-    "sourcegraph": "^18.4.0",
+    "sourcegraph": "^23.0.1",
     "tslint": "^5.11.0",
-    "typescript": "^3.1.4"
+    "typescript": "^3.7.3"
+  },
+  "dependencies": {
+    "@sourcegraph/prettierrc": "^3.0.1",
+    "rxjs": "^6.5.3"
   }
 }
